+++ eyes_selenium/applitools/selenium/visual_grid/__init__.py
@@ -1,6 +1,6 @@
 from .eyes_connector import EyesConnector
 from .render_task import RenderTask
-from .resource_cache import ResourceCache
+from .resource_cache import PutCache, ResourceCache
 from .resource_collection_task import ResourceCollectionTask
 from .running_test import RunningTest
 from .vg_task import VGTask
@@ -9,6 +9,7 @@ from .visual_grid_runner import VisualGridRunner
 
 __all__ = (
     "VGTask",
+    "PutCache",
     "RenderTask",
     "ResourceCache",
     "VisualGridEyes",