+++ tests/functional/eyes_selenium/desktop/test_fluent_api.py
@@ -227,14 +227,40 @@ def test_accessibility_regions(eyes_opened, check_test_result):
                 "actual_name": "accessibility",
                 "expected": [
                     AccessibilityRegion(
-                        10, 284, 800, 500, AccessibilityRegionType.LargeText
+                        10, 286, 800, 500, AccessibilityRegionType.LargeText
                     ),
                     AccessibilityRegion(
-                        122, 928, 456, 306, AccessibilityRegionType.LargeText
+                        122, 933, 456, 306, AccessibilityRegionType.LargeText
                     ),
                     AccessibilityRegion(
-                        8, 1270, 690, 206, AccessibilityRegionType.LargeText
+                        8, 1277, 690, 206, AccessibilityRegionType.LargeText
                     ),
+                ],
+            }
+        ]
+    )
+
+
+def test_check_window_with_match_region_paddings__fluent(
+    eyes_opened, check_test_result
+):
+    eyes_opened.check(
+        "Fluent - Window with ignore region by selector stretched",
+        Target.window()
+        .fully()
+        .ignore(".ignore", padding=dict(left=10))
+        .content("#stretched", padding=dict(top=10))
+        .layout("#centered", padding=dict(top=10, right=50))
+        .strict("overflowing-div", padding=dict(bottom=100)),
+    )
+    check_test_result.send(
+        [
+            {
+                "actual_name": "ignore",
+                "expected": [
+                    Region(10 + 10, 286, 800, 500),
+                    Region(122 + 10, 933, 456, 306),
+                    Region(8 + 10, 1277, 690, 206),
                 ],
             }
         ]